requirements.txt
requests
beautifulsoup4
python-vlc
yt-dlp
ibm-watsonx-ai
dotenv